add parse_response fuction, makes parse mod private
